fix type money

+++ db/migrate/20170910193200_create_users.rb
@@ -3,7 +3,7 @@ class CreateUsers < ActiveRecord::Migration[5.1]
     create_table :users do |t|
       t.string :firstName
       t.string :lastName
-      t.float :money            ,default: 0
+      t.decimal :money            ,default: 0
       t.string :email
       t.string :password_digest
       t.string   :confirmation_token
